add formatPhoneNumber
