+++ src/components/modals/RegistrationFormModal/FormMentor/formMentorScheme.js
@@ -1,4 +1,5 @@
 import { patternEmail, patternName, patternPhone, patternUrlLinkedin } from "@/src/constants/regulars";
+import { formatPhoneNumber } from "@/src/lib/utils/formatPhoneNumber";
 import { z } from "zod";
 
 export const mentorDefaultValues= {
@@ -14,59 +15,18 @@ export const mentorDefaultValues= {
 }
  
 // export function formatPhoneNumber(number) {
-// if( /^\+380[\s]\d{2}[\s]\d{3}[\s]\d{2}[\s]\d{2}$/.test(number)){return number}
+//   // Видаляємо будь-які нецифрові символи з номеру телефону
+//  const phoneNumber = number.replace(/\D/g, '');
 
-//   let formattedNumber = '';
-//   let numberString = number;
-//   formattedNumber += numberString.slice(0, 4) + ' ';
-//   formattedNumber += numberString.slice(4, 6) + ' ';
-//   formattedNumber += numberString.slice(6, 9) + ' ';
-//   formattedNumber += numberString.slice(9, 11) + ' ';
-//   formattedNumber += numberString.slice(11);
-//   return formattedNumber;
+//   // Перевіряємо, чи довжина номеру телефону дорівнює 12 (припускаємо, що вхідний формат правильний)
+//   if (phoneNumber.length === 12) {
+//       // Форматуємо номер у вигляді +380 67 760 65 08
+//       return `+${phoneNumber.slice(0, 3)} ${phoneNumber.slice(3, 5)} ${phoneNumber.slice(5, 8)} ${phoneNumber.slice(8, 10)} ${phoneNumber.slice(10, 12)}`;
+//   } else {
+//       // Якщо номер не відповідає очікуваній довжині, повертаємо не змінений рядок або можемо повернути помилку
+//       return number;
+//   }
 // }
-
-export function formatPhoneNumber(number) {
-  // Видаляємо будь-які нецифрові символи з номеру телефону
- const phoneNumber = number.replace(/\D/g, '');
-
-  // Перевіряємо, чи довжина номеру телефону дорівнює 12 (припускаємо, що вхідний формат правильний)
-  if (phoneNumber.length === 12) {
-      // Форматуємо номер у вигляді +380 67 760 65 08
-      return `+${phoneNumber.slice(0, 3)} ${phoneNumber.slice(3, 5)} ${phoneNumber.slice(5, 8)} ${phoneNumber.slice(8, 10)} ${phoneNumber.slice(10, 12)}`;
-  } else {
-      // Якщо номер не відповідає очікуваній довжині, повертаємо не змінений рядок або можемо повернути помилку
-      return number;
-  }
-}
-
-// Приклад використання:
-const phoneNumber = "=38067 760650 8";
-const formattedPhoneNumber = formatPhoneNumber(phoneNumber);
-console.log(formattedPhoneNumber); // Виведе: +380 67 760 65 08
-
-
-// const phoneFormat = (s) => {
-//   //const startsWith = plus ? '+7' : '8';
-//   //const startsWith = '+';
-//   //let x = parseInt(s.replace(/[^\d]/g, ''))
-//   let phone = parseInt(s.replace(/[^\d]/g, ''))
-// //let phone = s.replace(/[^0-9]/g, '');
-//   // if (phone.startsWith('7') && plus) {
-//   //   phone = phone.substr(1);
-//   // }
-//   // if (phone.startsWith('8')) {
-//   //   phone = phone.substr(1);
-//   // }
-//   return phone
-//   //return phone.replace(/\d{3}\d{3}\d{2}\d{2} /g, `${startsWith} $1 $2 $3 $4`);
-// };
-// console.log(phoneFormat('+380 67 760 65 08'))
- //89992223355 -> +7 999 222 33 55
-//  @param {string} s номер телефона
-//  @param {boolean} [plus=true] формат +7 или 8
-
-
 
 export const MentorSchema = z
 	.object({
